[model] make serde optional
Make the serde dependencies optional, allowing users to switch between
enabling and disabling them. This is done by a new feature,
`serde-support`, which comes with `serde`, `serde_json`,
`serde-mappable-seq`, and `serde_repr`. By default this feature is
enabled, but can be disabled like so:

```toml
[dependencies]
dawn-model = { default-features = false, git = "https://github.com/dawn-rs/dawn" }
```

Enabling only one or two of the optional serde dependencies (disabling
`serde-support`, but enabling `serde` and `serde_json` is not supported
and will do nothing. All of the dependencies are needed for correct
(de)serialization.

This has been documented in the readme.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>