[cache] add initial in-memory implementation
Add the initial in-memory implementation of the Cache.

The Cache in the core crates is split into three crates:

- `cache-trait`, which exports the fundamental traits and types needed
to build a working Cache implementation

- `cache-inmemory`, which is an implementation of a process in-memory
cache exposing no locks to userland

- `cache`, which re-exports the types from `cache-trait` and re-exports
the `cache-inmemory` crate (the in-memory crate can be disabled).

Libraries with a Cache implementation should depend only on
`cache-trait` for all of the building blocks.

The inmemory cache isn't yet finished, but most of it is there. Some
event update logic is missing, and there may be inconsistencies in
updates and deletes which may cause an inconsistent cache. These need to
be looked through.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>