[http] break out request fields into structs
Break out each of the Request structs in the `request` module into,
generally, two structs: the request struct itself and a struct of fields
which can be serialized.

This removes the need for requests themselves to implement Serialize,
which required that a lot of the request structs' fields be marked as
`#[serde(skip)]`.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>