[ci] fix workspace use in actions
Workspaces don't support `--all-features` in commands like `cargo test`,
which is a silent error and ignored, but fails our CI. To fix this,
remove the flag.

Additionally, switch out the deprecated `--all` for the new
`--workspace` flag, and remove the `cargo check` action since it's
covered by the `cargo test` action.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>