[http] return request finalizers on all of client
On all of the HTTP client's methods, return a finalizer request struct,
rather than returning a Future resolving to the finalized type.

This allows us to add optional fields and settings to requests over time
as Discord and we add features.

Using these requests is the same syntax, `client.action(...).await`.

Closes #26.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>