Add Cluster and redesign API
Closes #24

This commit adds support for the Cluster, which will collect together
multiple shards and and put all of them together into a single stream.

Furthermore this commit also changes the api multiple places from
returning a struct that implements `Stream<Item = Event>` to the type:
`impl Stream<Item = Event>`, this happens because of multiple reasons,
it makes the API more clear and makes it possible for us to change how
it works while still having the same API more or less. A downside of
this about it is that it limits our ability to make public methods on
the returned structs.

Signed-off-by: Valdemar Erk <valdemar@erk.io>