Added proper support for resume.
This change also led to a rewamp of some of the internal
as there was Arc that was getting invalidated, in a way
that no calls would work through it, so I had to change
some of the underlying structure to get the new session
up the call stack.

This is done with having a watch channel, a channel
that only hold the last value sent to it, and then updating
it to the new arc on every reconnect.

Signed-off-by: Valdemar Erk <valdemar@erk.io>