Make the queue trait more specific and support for large sharding.
This commit changes the queue trait so it can be used to support large
bots. It does that by adding a parameter to request that is the id of
the shard that you are requesting a identify on.

It also adds a new queue that is targeted large bots but can be used
with bot of all sizes if you set the queue size to be 1. This queue
also does bookkeeping so you do not go above the per day limit of
identifies.

Signed-off-by: Valdemar Erk <valdemar@erk.io>