Add new feature `inner-event-serialize`
The new feature will make the `Event` enum serializable.
It will be untagged though and only the data payload will
be in the exported JSON. Thus to deserialize it you may need
to use the other new method `event_name` that gives the name
of the event as discord sends it.

Signed-off-by: Valdemar Erk <valdemar@erk.io>