add embed validation
Add embed validation for creating and updating messages. This checks the
the author name, description, field names, field values, footer text,
and title lengths to ensure they aren't too long. The total number of
fields is checked as well as the total combined length of all of the
above.

This addresses comment
<https://github.com/twilight-rs/twilight/pull/167#issuecomment-633817770>.

Limits are referenced in the documentation to point to
<https://discord.com/developers/docs/resources/channel#embed-limits>.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>