model/gateway: export reaction_remove_emoji (#174)
In the gateway's payload module, export the `reaction_remove_emoji`
module. The `payload` module already re-exports the
`ReactionRemoveEmoji` struct within it, but that struct contains a
`PartialEmoji` local to its module. The `PartialEmoji` wasn't
re-exported, so the type couldn't be named.

`PartialEmoji` isn't important enough to export from the `payload`
module, so this exports the `reaction_remove_emoji` module.
Also, this fixes a type in `PartialEmoji`'s name from `ParitalEmoji`.

Approved-by: Arzte
Approved-by: Erk-
Merged-by: Vivian Hellyer <vivian@hellyer.dev>
Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>