gateway,http: fix stable clippy
Fix stable clippy which introduced a regression into stable that gives
this error message for all awaits:

```
warning: used binding `_task_context` which is prefixed with an
underscore. A leading underscore signals that a binding will not be
used.
```

See clippy issue 5360.

Additionally, allow an unrelated clippy lint that's incorrect.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>