standby: fix non-instant bystander waiting (#197)
When iterating over the bystanders of an event being waited on, if the
check didn't pass the sender that was taken from the bystander would
never be put back. This means that the next time the bystander was going
to be checked, the Standby struct would see it had no senders and would
remove it. This means that if a check wasn't immediately satisfied on
the first check that it would never successfully resolve and be removed.

A test has been included for supporting non-sequentual bystander
completions.

Approved-by: AEnterprise <aenterprise@aenterprise.info>
Approved-by: Valdemar Erk <valdemar@erk.io>
Merged-by: Valdemar Erk <valdemar@erk.io>
Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>