model: make member guild ids always present
This PR is another try at #180, which had to be reverted in #185 due to
not supporting deserialising guilds from the gateway. This adds support
by manually implementing Deserialize for Guild, and in the case of a
members key deserializing each element via the MemberDeserializer.

Change the `model::guild::Member`'s `guild_id` field from an
`Option<GuildId>` to just a `GuildId`. The reason that this was an
option before is because this was a naive serde-derived
(de)serialization implementation. The HTTP API doesn't return members
with guild IDs, but the gateway does. This made working with members
from the gateway awkward since you know the guild ID is there but had to
go through an extra pattern match to get it.

We can fix this by making a separate seeded deserializer implementation
if you already know the guild ID. The deserializer will just deserialize
into an intermediary member struct and create a real member with the
stored guild ID.

The `http` crate has been updated to do these deserializations and other
crates are updated to work with a guild ID that's always there.

Closes #42.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>