model/channel: fix guild text channel deserialization
Fix the deserialization of test channels by re-ordering the variants of
the GuildChannel enum. Since the enum is deserialized via
`#[serde(untagged)]`, the variants are tried in order for
deserialization. This means that, since a category channel shares all of
its fields with text channels, all text channels will deserialize into
category channels if the category channel variant is first.

Fix this by reordering the enum variants to put categories last and
voice channels first. Tests are included for deserializing every
variant.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>