gateway: separate booting of shards from creation
Separate the booting of a Shard (connecting to the gateway, aka
"running") from the initialisation of the struct.

This means that this code:

```rust
let shard = Shard::new(config).await?;
```

is now:

```rust
let mut shard = Shard::new(config);
shard.start().await?;
```

This is useful for separating the concerns of state initialisation and
the processing of an event loop.

Additionally, in the Cluster, instead of creating shards in
`Cluster::up`, create the shards on initialisation in the `Cluster::new`
method, which is now an async function due to being able to
automatically retrieve the total shard count from the gateway.

This means that cluster code may look like:

```rust
let cluster = Cluster::new(config).await?;
let mut events = cluster.events().await;

let cluster_spawn = cluster.clone();
tokio::spawn(async move {
    cluster_spawn.up().await;
});

while let Some(event) = events.next().await {
    // ...
}
```

Additionally, this means that the stream of events can now be retrieved
before starting and waiting on the cluster to boot shards, so events can
be received and processed while the cluster is still booting in the
background.

This is a successor to PR #148.

Closes #153.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>