gateway/shard: handle invalid authorization
In the gateway's Shard, handle invalid authorization both early and
during processing of the shard processor's loop.

While creating the shard, get the gateway URL from the HTTP API prior to
creating the processor. This allows an early return if getting the
gateway URL (such as due to invalid authorization) failed. Additionally,
during the processing of the shard processor's loop, if a close frame
with a 4004 error (invalid authorization) is encountered, remove all of
the listeners for the shard and exit the loop.

Closes #152.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>