standby: use dashmap, return future structs
In Standby, use DashMap for internal maps, which avoids the need for
locking and many of the functions needing to be asynchronous.

The waiting methods, such as `Standby::wait_for`, now return futures
which resolve to the same return type that they did before. "Voldemort
types", or types that can't be named (such as async fns, or `impl
Future`s), are sub-par due to needing to be boxed when stored somewhere.
By having named structs, this can make handling futures more efficient
for users.

This PR could have been separated into two, but that would have nearly
doubled work and would end up in conflicts between the two, so it was
simpler to combine them together.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>