cache/in-memory: impl channel pins update event (#212)
Implement handling of the ChannelPinsUpdate event. This will first look
through the guild channels for a channel by the ID and, if it's a text
channel, will update the last pin timestamp. If a guild channel by the
ID doesn't exist, it will look through the private channels, and lastly
the groups.

This was the last unimplemented event, so this closes #65.

Approved-by: Arzte
Approved-by: Valdemar Erk <valdemar@erk.io>
Merged-by: Vivian Hellyer <vivian@hellyer.dev>
Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>