gateway/cluster: fix command dispatching (#232)
When the cluster would start a shard, it clones the shard to reduce the
time it holds on to the map. The shard didn't hold everything behind an
Arc, so some of the data would be cloned and, in *that new clone of the
shard*, would be modified. This means that the shard temporarily cloned
by the cluster would be in a different state from that of the one stored
in the cluster's internal map.

To solve this, put everything behind an Arc. Since we need to lazily
initiate some values of the shard after initialization, put it behind a
OnceCell.

Closes #218.

Approved-by: AEnterprise <aenterprise@aenterprise.info>
Approved-by: Valdemar Erk <valdemar@erk.io>
Merged-by: Vivian Hellyer <vivian@hellyer.dev>
Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>