model: remove 'serde-support' feature  (#243)
Remove the optional 'serde-support' feature in the model crate. As
evidenced by this patch, this feature adds a lot of complexity to the
crate. Most users of the twilight ecosystem are, somewhere in their
tree, going to be using serde. There will be very few use cases, such as
some specific cache layers, that won't need to use serde, but these use
cases are typically few and far between, and a couple of runtime
dependencies won't hurt much.

All tests have essentially not been touched, documentation for the
feature has been removed, and other crates have been updated to not
mention the feature.

Approved-by: Arzte
Approved-by: Valdemar Erk <valdemar@erk.io>
Merged-by: Vivian Hellyer <vivian@hellyer.dev>
Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>