model/event/gateway: optimise deserialiser
Optimise the deserialisation of
`twilight_model::gateway::event::GatewayEvent` by scanning the input
JSON payload to deserialise prior to actually deserialising it.

Instead of implementing `serde::Deserialize` directly on `GatewayEvent`,
make a `GatewayEventDeserializer` that implements
`serde::de::DeserializeSeed` and outputs a `GatewayEvent` on success.

When making a `GatewayEventDeserializer`, it will scan the payload for
an "op" and parse out the opcode value, and optionally a "t", which
denotes the event type in the case that the gateway event is a Dispatch
event (op 0).

When deserialising the gateway event itself, the op can be matched on
and logic can be applied based on the opcode. For example, for the event
op the sequence will be parsed and the dispatch payload will be parsed
due to the event type already being known. For a heartbeat opcode, only
the sequence will be parsed, and other fields can be ignored.

This removes the use of an intermediary `serde_value::Value` that the
dispatch payload (such as an entire guild create's payload, or a role
create payload, etc.) was being deserialised into before later being
deserialised into the destination type.

This improves the performance of a simple `GUILD_ROLE_DELETE` dispatch
event's deserialisation by 30%:

```
gateway event role delete
                        time:   [460.83 ns 463.49 ns 467.09 ns]
                        change: [-30.530% -30.153% -29.713%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
```

An approximately 1000 member guild now sees a 62.3% improvement in
performance:

```
gateway event guild create
                        time:   [1.9464 ms 1.9503 ms 1.9544 ms]
                        change: [-62.539% -62.301% -62.080%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>