model/channel: add a few convenience methods
Add a few convenience methods to
`twilight_model::channel::{Channel, GuildChannel}`:

- `Channel::id`
- `Channel::name`
- `GuildChannel::guild_id`
- `GuildChannel::id`
- `GuildChannel::name`

These are a few operations that are done a lot, probably by most users.
All channels and guild channels have an ID and can have a name, so both
have matching methods for retrieving both from the inner variants.
`GuildChannel` is the only one with a `guild_id` method since it only
applies to guild channels, and not other types of channels.

Closes #217.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>