project: add rusty-hook
Add `rusty-hook` to the development dependencies of the main `twilight`
crate. When `cargo test` is first run, it will automatically add a
pre-commit hook. This patch configures the pre-commit hook to run cargo
fmt, cargo clippy, and then cargo test.

This will help ensure that (probably) properly building PRs are pushed
the on every PR commit and reduce iteration.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>