lavalink/node: ignore when user drops rx
On every iteration of the node connection loop a branch would run to
check if the user hadn't dropped the rx of events received from
lavalink, and in the case that they did, break the loop. Not all users
want to handle these, so just ignoring if the user closed the channel is
fine.

Closes #296.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>