project: revamp embed builder (#361)
* project: revamp embed builder

The `twilight-builders` crate is a bit odd: it has no documentation and
only contains one thing, a module named `embed`. The embed builder in
this crate is even more odd, and uses a consuming and "commit" style
builder that's uncommon in the Rust ecosystem, or possibly not a pattern
used in any other crate.

Twilight doesn't rely on a common set of builders across the ecosystem,
which is why this is the only builder. There likely won't be a need for
another builder, and if there is then it should be localised to a more
specific crate, such as the gateway or HTTP.

This patch renames the `twilight-builders` crate to
`twilight-embed-builder` and revamps the embed builder. The embed
builder is now consuming, and validates arguments. For example, creating
a field builder will validate that the input name and value is not empty
or too long, or that an author embed's name is not too long. If they
are, a localised error is returned. This means that the crate doesn't
share one enum for every operation, and that each method has its own
error enum so that all variants are always relevant to the operation.

When an embed builder is built into the final product - an embed - it
checks that the number of fields is not too many and that the total
textual length of the embed, counting things like the field names and
values and footer text, is not too long and would not be rejected by
Discord.

Extensive documentation has been added and examples have been added for
many of the methods and types.

This is part of #239.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* not plural

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* add docs

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* add readme sync workflow

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* move to utils folder

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* re-export embed footer text error

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* add image source abstraction

Add an abstraction over attachment and URL image sources, which each
individually do checks to make sure they are valid.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* update doc workflow directory

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* update embed builder readme

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* fix image source url error variant anchor name

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* make thumbnail example use thumbnail

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* add link to embed builder in twilight readme

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* add author name checks

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* add more docs

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* abstract field max length values

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* add some tests

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* add to readme

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>
Approved-by: Valdemar Erk <valdemar@erk.io>
Approved-by: Cassandra McCarthy <cassie@7596ff.com>
Merged-by: Cassandra McCarthy <cassie@7596ff.com>