gateway: refactor event emissions
Refactor event emissions to be simpler, faster, reduce allocations, and
remove duplicate code.

Instead of creating a `Vec` to store listener IDs to remove, iterating
over listeners and determining what to remove, and then iterating over
the vec to remove them, use `DashMap::retain` to directly remove
listeners that have had their channels closed.

Additionally, simplify and remove code by abstracting the iteration and
checks over the listeners map for the `bytes` and `event` emission
functions to another function that both call. `bytes` and `event`
provide a function to create events, and this function is called when
sending an event to a listener channel.

`bytes` now takes a reference to the listeners map and is no longer an
async fn.

Additionally, add documentation and tests.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>