cache,gateway,http: remove tracing imports (#336)
Remove the imports for `tracing` crate macros. In some cases, these are
only needed when a feature is enabled, so we need to put
`#[allow(unused)]` on them, but in a couple of cases these imports
existed despite no longer being ever used. For consistency with the rest
of the libraries and in general, don't import them and just call them
via FQN.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>