gateway: refactor error handling  (#376)
* gateway: refactor error handling

Refactor error handling in the gateway crate, including internal errors
and public API errors.

The gateway crate contained 3 error enums: the cluster error, shard
error, and shard processor error. This means that, for example, all
cluster operations, ranging from creating a cluster configuration to
sending a command to a cluster's shard, would return a cluster error
even if only one or two were applicable to a function.

This becomes more of an issue when working with shard operations:
retrieving the latency information of a shard would return an error enum
with 8 potential variants. In reality, there's only one reason doing
this would fail.

This makes handling errors confusing, and creates a need to reference
the documentation for every function to know which variants apply. This
can be solved by returning errors specific to each operation.

One example of the new approach is that, when a shard is started, a
`ShardStartError` enum is returned containing three variants:
`RetrievingGatewayUrl` when retrieving the URL over HTTP fails,
`ParsingGatewayUrl` when parsing the gateway URL fails, and
`Establishing` when establishing the WebSocket connection fails.
Determining and handling why starting the shard failed is clear and
consise.

A second example is setting the large threshold for a shard. Instead of
returning a general shard error with a *lot* of irrelevant variants, a
specialized `LargeThresholdError` enum is returned with two variants:
`TooFew` when too few members in the threshold is provided and `TooMany`
when too many members are in the threshold.

All of these enums implement `Display` and `Error`, making handling them
with an error handling crate easy.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* rename shard start url variant

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* line length

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* fix simd json error import

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* actually test it this time

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* address comments

Approved-by: BlackHoleFox <blackholefoxdev@gmail.com>
Approved-by: Cassandra McCarthy <cassie@7596ff.com>
Merged-by: Cassandra McCarthy <cassie@7596ff.com>
Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>