gateway: make much of the public api sync (#382)
Make much of the Cluster and Shard API synchronous. A lot of the
functionality, mainly surrounding the usage of asynchronous Mutex locks,
doesn't actually need to be asynchronous. The main benefit of async
locks is when they're going to be held across context points, i.e. when
an `await` is encountered. The running task may be switched to another
on the same thread and, if acquiring the same lock, will deadlock. Since
this can't happen due to the locks not being used in asynchronous
contexts, there's no need for them, thus removing the need for their
caller functions to be async.

This makes the following API surface sync:

- `Cluster::down`
- `Cluster::down_resumable`
- `Cluster::events`
- `Cluster::info`
- `Cluster::shard`
- `Cluster::some_events`
- `Shard::events`
- `Shard::info`
- `Shard::some_events`
- `Shard::shutdown`
- `Shard::shutdown_resumable`

The following Cluster and Shard API surface is still async:

- `Cluster::command` (due to ratelimiting)
- `Cluster::new` (due to retrieving bot user session availability)
- `Cluster::up` (waiting for the shards to be up)
- `Shard::command` (same as `Cluster::command`)
- `Shard::start` (same as `Cluster::up`)

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>