model: handle new permissions scheme
Permissions in Permission Overwrites and Roles now have two fields in
the API: bitflags stored as an integer and another as a string.

Permission Overwrites now have the following fields:

- allow_new
- deny_new

Roles now have:

- permissions_new

For the motivations behind this change read the link in #365. This
patch introduces support for these fields. To make a migration easier
(or, requiring no effort), rename the Permission Overwrites' `allow` and
`deny` fields to `allow_old` and `deny_old`, adding `allow_new` and
`deny_new` as `allow` and `deny`. Additionally, do the same for Role's
`permissions`: rename it to `permissions_old` and make the new one
`permissions`. This can be done simply via `#[serde(rename)]`.

Tests for permissions, permission overwrites, and roles have been added.

Closes #365.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>