gateway/shard: fix command send ratelimiter
Fix the ratelimiter for sending events over the shard. The ratelimiter
uses tokio's interval, which like documented in #268 and the comments of
issue #193, isn't what you'd think it is at first glance.

Tokio's Interval will create an unbounded number of "tickets" available
every specified duration, adding a new one each time that the duration
has passed. This means that if you have an interval that refreshes every
5 seconds and take from it twice, it'll take 5 seconds to do in total.
If you wait a few minutes and then take from it 10 times, all of those
tickets will be immediately retrieved. This isn't what you want with a
ratelimit.

Instead, use Tokio's Throttle. It's similar but has a differing detail:
it will only keep 1 ticket available at a time. This means that if you
wait a few minutes, take a ticket, and then try to immediately take
another, unlike with the Interval you'll have to wait, since it counts
down from when the *last* ticket was taken.

Closes #193.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>