project: make readme badges consistent
Make the badges consistent across all of the crate READMEs (and slightly
prettier). Some of them have different versions, and most out of them
are no longer true. For example, the model crate requires 1.31+ (which
is probably not be true), the command parser requires 1.36+, and the
main crate requires 1.40+.

Our policy for now is to support latest stable, so instead all of these
have been updated to mention "stable". Some of the READMEs don't have
badges, to add them to those READMEs.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>