gateway: rework cluster and shard configuring (#422)
Rework Cluster and Shard configuring to be a bit simpler and require
working with fewer types.

Both the Cluster and the Shard allow you to create them by just
providing the token to use. Configuring them is a bit more complex: you
have to use a `ClusterConfigBuilder` to construct a `ClusterConfig` and
then pass the configuration to `Cluster::new`. The same is true for the
shard.

It's simpler to have only two types that need to be worked with,
`ClusterBuilder` and `Cluster` (and the shard equivalents). Instead of
using the builder to construct a configuration, use it to construct the
matching struct directly.

This reworks the code required to configure and construct a Cluster
from:

```rust
use twilight_gateway::cluster::{ClusterConfig, Cluster, ShardScheme};
use twilight_model::gateway::GatewayIntents;

let config = ClusterConfig::builder(&token)
    .shard_scheme(ShardScheme::Auto)
    .intents(Some(GatewayIntents::GUILD_MESSAGES))
    .build();

let cluster = Cluster::new(config).await?;
```

to a more concise:

```rust
use twilight_gateway::cluster::{Cluster, ShardScheme};

let cluster = Cluster::builder()
    .shard_scheme(ShardScheme::Auto)
    .intents(Some(GatewayIntents::GUILD_MESSAGES))
    .build()
    .await?;
```

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>