cache/in_memory: de-asyncify almost all interfaces (#421)
In the InMemoryCache, make all of the async functions non-async except
for the CacheUpdate implementations due to trait requirements. Since we
layer over DashMap, we don't need these to be async.

The only other logic change is the current user: instead of using a
futures-util Mutex to store the current user, just use a Mutex from the
stdlib. So long as the lock isn't held across await points and no panic
points are accessible during the guards' lifetime the lock won't poison.

Additionally, make all of the `InMemoryCache` methods not return
results. The methods can never return an error, so there's no point to
doing this.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>