cache: rework crates (#427)
Rework the cache crate setup to simplify down from 3 (4, kind of) crates
to just 1 crate.

There are currently 4 crates:

- base
- in-memory
- redis
- trait

The trait crate was meant to implement a foundation for creating caches,
but it was never finished and an issue is open to completely rework it.
As a result, this doesn't really do anything. Its functionality is being
merged with the in-memory crate.

The redis crate was never worked on and is empty, so it's being removed.
An issue to make one is open.

The base crate re-exported the types from the trait crate and the
`InMemoryCache` type from the in-memory crate. Since the traits aren't
useful for the user, it doesn't need to be exported. This makes the base
crate only re-export the `InMemoryCache`. Instead, just remove the crate
and only have the `cache-inmemory` crate for now. Documentation has been
updated, the `update` function is no longer async or return results with
empty errors (since it can't error), and the documentation for the main
`twilight` crate has been updated.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>