gateway/inflater: refactor internals (#419)
Refactor the internals of the inflater to be clearer and to split a few
more pieces of functionality into their own functions. Make the tracing
logs clearer and utilise the functionality of tracing to a greater
extent.

One small logic change has been applied: the logic to check if the
buffer length is at least 1/4 the size of the capacity was incorrect.
The math was basically `len > capacity * 4`. This can clearly never be
the case, so the buffers were never being shrunk. This has instead been
changed to `len > capacity / 4`.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>