cache/in-memory: make unnecessary methods private
Hide unnecessarily public methods in the `InMemoryCache` that should
only be used internally. These were made public in #141 on the premise
of users using the cache to cache objects over HTTP, but this introduces
race conditions and isn't the intent of the cache, which is meant to be
used for sequential gateway events.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>