gateway: conditionally deserialize received events
In the gateway's shard processor, conditionally deserialize received
events if the user has specified that they want them.

Intents are a good way of filtering events that a user wants, but it
only filters by category. If a user selects the `GUILD_MESSAGES` event
then they will receive `MESSAGE_CREATE`s, `MESSAGE_DELETE`s,
`MESSAGE_UPDATE`s, and so on. Intents are especially good at reducing
network usage, but can cause users to still receive and deserialize more
events than they actually use.

Say that a user wants to handle `MESSAGE_CREATE`s but not
`MESSAGE_DELETE`s. They don't care if message deletes are deserialized
or not since they don't use them. Coincidentally, shards don't use most
dispatch events either, such as message deletes. We can utilize event
filtering through the `Cluster::some_events` and `Shard::some_events`
methods to determine whether a message should be deserialized or not. If
not, then we can just receive the payload over the websocket and
immediately forget about it.

Shards care about only a specific handful of events: the Ready and
Resumed dispatch events and other non-dispatch gateway events. For the
Resumed event, we don't actually need to go through a serde context to
deserialize it and can construct it and process it. This is the same for
the Heartbeat Acknowledgement gateway event. The Ready event is another
that shards need to process and contains a lot of information. We can
easily use serde to deserialize the only field of Ready events that we
need.

For some use cases this may only see a few percentage points in
performance difference, but for others this may be up to 50% depending
on what the application does. Notes have been added to
`Cluster::events` and `Shard::events` that it is highly recommended
users use their `some_events` counterpart to utilize this new event
deserialization efficiency.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>