gateway/examples/metrics: reduce features
In the gateway's "metrics" example, reduce the features of dependencies,
mainly `tracing-log` and `metrics-runtime`. `metrics-runtime` in
particular by default brings in all observers and exporters, such as
HTTP exporters and prometheus observers.

This reduces the example's dependency count from 232 to 216 and the
workspace's total dependency count in a build from 262 to 256.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>