project: turn main crate into advertisement crate
Turn the "main crate" - `twilight` - into an advertisement crate that
documents the existence and usage of other crates, linking to their
documentation on the `twilight.rs` website. The crate now has no
contents, dependencies, or features, and has a warning in the first few
sentences that is has no functionality. The installation instructions
now specifically recommend installing the individual crates and provides
a sample list of `Cargo.toml` `[dependencies]` entries that are likely
to be popularly used.

This is a part of #459.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>