gateway/shard/processor: deserialize ready once (#492)
Deserialize the ready event only once. Ready events were being specially
handled and deserialized only partially by the shard to avoid
unnecessarily deserializing the rest if the user wants them. When using
`simd-json` it will modify the input, rendering it unusable when
deserializing a second time when emitting the full ready. This
deserializes it as a full ready event every time, making it deserialized
only once.

Closes #491.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>