mention: add parsing functionality (#513)
* mention: add parsing functionality

Add parsing functionality to the `twilight-mention` crate to pair with
its existing formatting capability.

There are two key pieces here: `MentionParse`, which is a trait defining
`parse` and `iter` methods, and `MentionIter`, which will iterate over a
buffer and parse mentions out of it.

`MentionParse` contains 4 members: `PARTS`, `SIGIL`, `parse`, and
`iter`.

`SIGIL` defines the "sigil" of a mention; in the case of user mentions,
the sigil is "@", and in the case of emojis it is ":".

`PARTS` defines the number of parts are separated by the sigil. User
mentions have one part and emojis have two parts (as they're separated
by the sigil ":").

`parse` is a method that parses a mention out of a buffer, where the
buffer have no extraneous characters. This means it will parse the
buffer "<@123>" as a user ID of 123, but error on the buffer "user
<@123>".

`iter` is a method that will initialize an iterator over a buffer to
find and parse mentions. For example, in the string "the channels <#123>
and <#456>" will create an iterator of `twilight_model::id::ChannelId`
with values 123 and 456. Unlike `parse`, iterators don't produce errors.

`MentionParse` is implemented for `twilight_model::id`'s `ChannelId`,
`EmojiId`, `RoleId`, and `UserId`.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* iter: include start and ending indexes

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* misc refactoring

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* handle user nickname mentions

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* add 'mention type' wrapping enum

Add a `MentionType` enum to wrap all possible types of mentions. This is
useful when needing to, say, iterate over any possible mention in a
buffer.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* remove parsemention::sigils

Remove `ParseMention::SIGILS` because it's no longer needed.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* add validation note

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>