http: add auditlogreason trait (#522)
Add an `AuditLogReason` trait and implement it for all requests that
are auditable with reasons and will result in audit log entries. The
trait includes a `reason` method that will validate its input to
ensure it's valid. This trait is useful for generalizing reason
setting.

All requests' individual `reason` methods are now deprecated in favor of
the trait's interface.

Signed-off-by: Valdemar Erk <valdemar@erk.io>