gateway: bump to v0.2.0-beta.0
This beta version of major version 0.2 of the crate includes changes
needed to support version 8 of the Discord Gateway API.

Changes

`twilight-gateway` now depends on `twilight-http` 0.2 and
`twilight-model` 0.2.

`twilight_model::gateway::Intents` is now re-exported as
`twilight_gateway::Intents`.

The following methods now take a second "intents" parameter, as this is
now required to be specified by the API:
- `cluster::ClusterBuilder::new`
- `cluster::Cluster::builder`
- `cluster::Cluster::new`
- `shard::ShardBuilder::new`
- `shard::Shard::builder`
- `shard::Shard::new`

The `shard::Config::intents` method no longer returns an option and now
returns a copy of the intents (returning `twilight_gateway::Intents`).

[#532]: https://github.com/twilight-rs/twilight/pull/532

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>