Create a new CurrentUserGuild type which is a very stripped partial guild (#550)
The `GetCurrentUserGuilds` requests was failing because the partial
guild sent by Discord were more partial than the `PartialGuild` in the
model. This change-set includes a new type `CurrentUserGuilds` which
is a more stripped down version of the `PartialGuild`. The new type is
only used for this HTTP endpoint so it is placed in the twilight-http
crate.

Unit tests have been added for the new type to test the implementation

Closes #539

Co-authored-by: Vivian <vivian@hellyer.dev>
Co-authored-by: Jens Reidel <jens@troet.org>
(cherry picked from commit 18530e91e2ade75baec655af6fa551b049a83ec4)